bugfix(collection): Prevent instances from sharing collection proxy
Collection instances were sharing their proxy as a single closure
variable that each instance had access to, and the last instance would
overwrite. This PR adds a WeakMap to track collections per-instance
instead, preventing this issue. The map is still created per-collection
property, so there is no chance of overlap. This may be suboptimal and
in the future we should try to make it better (extended collections will
create an unused collection map, for instance) but should be fine for
now.