fix: depend on ember-auto-import alongside ember-cli-page-object
This addon is a transitive dependency of a Ember Addon I'm using in a
private repository.

Without this change, I receive a built-time error when I use the addon
in my Ember Engine project:
`ember-classy-page-object needs to depend on ember-auto-import in order
to use ember-cli-page-object`

Including ember-auto-import in the package.json dependencies section
of this addon fixes that error in my project.